+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 CHANGELOG
 =========
 
-1.7, 2022-07-23
+1.7, 2022-07-24
 ---------------
 - count a long keypress as a single event by default (issue #22)
 - add tray menu option to toggle sticky long keypress
+- add weekdays to statistics pages
 
 
 1.6, 2022-07-13